Fully Completed Sync Resetting System

+++ api/src/main/java/com/bgsoftware/superiorprison/api/data/mine/shop/MineShop.java
@@ -2,17 +2,18 @@ package com.bgsoftware.superiorprison.api.data.mine.shop;
 
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public interface MineShop {
 
     <T extends ShopItem> Set<T> getItems();
 
-    void addItem(ItemStack itemStack, double price);
+    void addItem(ItemStack itemStack, BigDecimal price);
 
     void removeItem(ShopItem item);
 
-    double getPrice(ItemStack itemStack);
+    BigDecimal getPrice(ItemStack itemStack);
 
     boolean hasItem(ItemStack itemStack);
 }
