+++ api/src/main/java/com/bgsoftware/superiorprison/api/data/mine/shop/ShopItem.java
@@ -2,9 +2,11 @@ package com.bgsoftware.superiorprison.api.data.mine.shop;
 
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
+
 public interface ShopItem {
 
-    double getPrice();
+    BigDecimal getPrice();
 
     ItemStack getItem();
